* Improves performance for  `LLAMA_CUDA_FORCE_DMMV=ON` on older GPUs. E.g., on GTX-1660 TG-128 time is 34.5 ms/t with this PR vs 37.1 ms/t on master. The change is performance neutral on modern cards (RTX-4080).
* For `LLAMA_CUDA_FORCE_DMMV=OFF`
  - Recover `Q4_K` performance on modern GPUs. E.g., on my RTX-4080 TG-128 time drops to 8.4 ms/t from 9.15 ms/t and is again comparable to `Q4_0` and `Q4_K` with `LLAMA_CUDA_FORCE_DMMV=ON` performance
  - On a somewhat older card (GTX-1660) TG-128 is reduced from 35.5 ms/t to 29.5 ms/t. This is better than the  `LLAMA_CUDA_FORCE_DMMV=ON` performance. 

I'm noticing that the `LLAMA_CUDA_FORCE_DMMV=OFF` CUDA version is completely broken for a k-quants block size of 64 (it does not even compile). I guess, I will be fixing this in a separate PR.

Oh, also fixed an annoying warning that `g_compute_capabilities` is not used, which is triggered when `LLAMA_CUDA_FORCE_DMMV=ON`