@JohannesGaessler 2023-07-22
>I'm noticing that the LLAMA_CUDA_FORCE_DMMV=OFF CUDA version is completely broken for a k-quants block size of 64 (it does not even compile). I guess, I will be fixing this in a separate PR.

Yes, I did not implement support for a k-quant block size of 64 because I was under the impression that this is only intended as a temporary fix anyways.