@JohannesGaessler 2023-07-22
The new kernels are faster on my systems:

| GPU      | Model     | Test  | t/s master | t/s PR | Speedup |
|----------|-----------|-------|------------|--------|---------|
| RTX 3090 | 7b q4_k_s | tg128 |     102.81 | 116.34 |    1.13 |
| RTX 3090 | 7b q4_k_m | tg128 |      98.28 | 111.54 |    1.13 |
| P40      | 7b q4_k_s | tg128 |      33.11 |  36.09 |    1.09 |
| P40      | 7b q4_k_m | tg128 |      32.00 |  34.00 |    1.06 |