@JohannesGaessler 2023-07-22
I tested `LLAMA_CUDA_FORCE_DMMV=ON` as well, the performance on my RTX 3090 is ~30% lower, on my P40 it's ~10% lower (than the `LLAMA_CUDA_FORCE_DMMV=OFF` performance with this PR).