@JohannesGaessler 2023-07-22
Actually I forgot about that option entirely so it should have been 2 as per default. Setting it to 1 reduced performance on both cards.